[updated] ai
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@3576 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/creature/AiAgentImplementation.cpp
@@ -262,13 +262,16 @@ void AiAgentImplementation::selectWeapon() {
 	}
 
 	ManagedReference<WeaponObject*> currentWeapon = getWeapon();
+	ManagedReference<WeaponObject*> defaultWeapon = dynamic_cast<WeaponObject*>(getSlottedObject("default_weapon"));
 
 	if (currentWeapon != finalWeap) {
-		if (currentWeapon != NULL)
+		if (currentWeapon != NULL && currentWeapon != defaultWeapon)
 			currentWeapon->destroyObjectFromWorld(false);
 
-		transferObject(finalWeap, 4, false);
-		broadcastObject(finalWeap, false);
+		if (finalWeap != defaultWeapon) {
+			transferObject(finalWeap, 4, false);
+			broadcastObject(finalWeap, false);
+		}
 	}
 
 	//setWeapon(finalWeap, true);
